Extract prompt-related functions
