+++ src/pages/lessons/components/LessonPrompts/hasWordBoundaryError.tsx
@@ -1,0 +1,6 @@
+export const hasWordBoundaryError = (
+  currentPhrase: string,
+  actualText: string
+) => {
+  return currentPhrase === "wars" && actualText.includes("Star Wars");
+};
